==> Capstone_1 (old)/1) Collecting_Data_TableDraft.ipynb <==
# Collecting Table Draft Data Notebook

This is a notebook aimed at collecting data from the Pro-Football-Reference website. Our goal is to create a table of WR's drafted, which includes data from their time in the NFL and at the college level. To do so, we scrape the draft table from the website called **draft_table**, which includes basic information about each player. Then, we scrape more data from the NFL and college webpages of each player and store it in the columns of each corresponding row. (step 0)

We then clean and verifiy the validity of the generated data. Issues come from the fact the webpages are missing or the NFL/college scrapped data isn't correct: hidden tables, disordered data, etc... In order to recognize how and where the additional data is generated, we also generate columns of comments for each player corresponding to messages that describe how the data was scrapped.

# Summary

We scrap the pro-football-reference website as follow. There are multiple sets of potential errors, so we clean the data one step at a time:

### 0) First, generate the data:
- tabledraft scrapes the draft table from the soup document and returns all players drafted with columns list_of_drafts: table_draft
- cfb_url_fill_in will generate the cfb url for those that are missing
- expand_draft_table adds the columns list_of_added_columns
- using the urls of table_draft, collegestats and nflstats both operate similarly and generate tables respectively for college football and the nfl
- row_cfb and row_nfl both pick the aggregate row from the provided table. It is capable of handling multiple cases of playing for many teams and not playing at all
- main puts it all together (this is the only command executed in this notebook). Watch as we also provide additional comments for each row: the caption from the nfl table and the method used. If numbers are generated it is a SUCCESS FROM MAIN, otherwise a FAIL (for example if the link is broken)

### 1) 1st set of cleaning:
- if the nfl table caption doesn't contain appropriate keywords, we must reconsider the nfl web page
- we use the webdriver.Chrome to find hidden tables
- as soon as we pick one with appropriate caption, we select it
- we re-run nflstats on it
- we edit the method used: SUCCESS FROM MAIN1. If the code failed to capture the appropriate table, write FAIL AFTER MAIN1

### 2) 2nd set of cleaning:
- if the yards from the original draft table ('receiving yards') don't match those from the nfl table ('nfl yards'), something is wrong. We identify those mismatches using identifier_for_main2 and print the URL of those nfl tables
- we manually inspect each nfl table. Very frequently, the mismatch rises because the nfl table has the rushing/receiving stats switched. For those, we run main2 which automatically does the flip. If the mismatch was taken care of this, the method reads SUCCESS FROM MAIN2, otherwise FAIL AFTER MAIN2

### 3) 3rd set of cleaning:
- we once again compare 'receiving yards'and 'nfl yards'. If they still don't match we print the url and work manually the nfl tables. We feed those mismatches and the list of manually edited nfl tables into main3

### Importing the module 
Collecting_Data_TableDraft_Functions.py (to be initialized)

import Collecting_Data_TableDraft_Functions
import importlib
importlib.reload(Collecting_Data_TableDraft_Functions)
from Collecting_Data_TableDraft_Functions import *

# Getting the draft table and Export

#url_wr_draft_1 = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=0'
#table_draft_1 = main(url_wr_draft_1)

#url_wr_draft_2 = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=300'
#table_draft_2 = main(url_wr_draft_2)

url_wr_draft_3 = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=600'
table_draft_3 = main(url_wr_draft_3)

#table_draft_1.to_csv(r'data_sets/wr_draft_1_complete.csv',index=False)
#table_draft_2.to_csv(r'data_sets/wr_draft_2_complete.csv',index=False)
table_draft_3.to_csv(r'data_sets/wr_draft_3_complete.csv',index=False)

td1_0 = pd.read_csv('data_sets/wr_draft_3_complete.csv')

## 1st Set of Cleaning: making sure wrong NFL tables are off

td1 = main1(td1_0)

td1.to_csv(r'data_sets/wr_draft_3_complete1.csv',index=False)

td1  = pd.read_csv('data_sets/wr_draft_3_complete1.csv')

#wrong_nfl = check_nfl_table(table_draft_1)
#display(table_draft_1.loc[still_wrong_nfl])
still_wrong_nfl = check_nfl_table(td1)

[print(x) for x in td1.loc[still_wrong_nfl,'nfl url']]

### Manual Editing:
most likely the still_wrong_nfl list is a set of players with no 'Receiving & Rushing' stats

#td1.columns.get_loc("nfl target")
for i in still_wrong_nfl:
    td1.loc[i,['nfl age','nfl pos', 'nfl no']] = np.nan*3
    td1.iloc[i,47:71] = np.nan*24
display(td1.loc[still_wrong_nfl])

## 2nd Set of cleaning: mismatch between NFL numbers

On the NFL table, the rushing stats and the receiving stats may be swapped. This issue often comes up because the NFL numbers from the draft table and from the NFL table don't match 

mismatch_yards = identifier_for_main2(td1)
[print(x, row ['receiving yards'], row['nfl yards'], row['nfl url']) for x, row in td1.loc[mismatch_yards].iterrows()]

recovered_mismatch = [ ]
td1 = main2(td1, recovered_mismatch, mismatch_yards)

mismatch_yards = identifier_for_main2(td1)
[print(x, row['nfl url']) for x, row in td1.loc[mismatch_yards].iterrows()]

td1.to_csv(r'data_sets/wr_draft_3_complete3.csv',index=False)

td1 = pd.read_csv('data_sets/wr_draft_3_complete2.csv')

## 3rd Set of Cleaning: Manual editing

#example of wrong table + correct table inverted

# url = 'https://www.pro-football-reference.com/players/W/WebbJo00.htm'
# url = 'https://www.pro-football-reference.com/players/L/LawrQu00.htm'
# url = 'https://www.pro-football-reference.com/players/S/SlatMa00.htm'
table_list = generate_all_nfl_tables(url)
table = find_nfl_table(table_list)
table3 = nflstats1(table)

td1 = main3(td1,[],[])

td1.to_csv(r'data_sets/wr_draft_3_complete3.csv',index=False)

# Trying out some particular cases in CFB

#all normal
marquise_brown_cfb_url = 'https://www.sports-reference.com/cfb/players/marquise-brown-1.html'

#multiple schools
jalen_hurd_cfb_url = 'https://www.sports-reference.com/cfb/players/jalen-hurd-1.html'

url_cfb_sample = marquise_brown_cfb_url
html =urlopen(url_cfb_sample)
soup = BeautifulSoup(html,'html.parser')

table=collegestats(soup)

display(table)
print(row_cfb(table))

# Trying out some particular cases in the NFL

#all normal
mike_thomas_nfl_url = 'https://www.pro-football-reference.com/players/T/ThomMi04.htm'

#no play time
jalen_hurd_nfl_url = 'https://www.pro-football-reference.com/players/H/HurdJa00.htm'

#little play time
juwann_winfree_nfl_url = 'https://www.pro-football-reference.com/players/W/WinfJu00.htm'

#played with multiple teams
odell_beckham_jr_nfl_url = 'https://www.pro-football-reference.com/players/B/BeckOd00.htm'

#Has a stat kick and punt returns
dwayne_harris_nfl_url = 'https://www.pro-football-reference.com/players/H/HarrDw00.htm'

#has nfl stats inverted (rushes before receiving)
ardarius_stewart_nfl_url = 'https://www.pro-football-reference.com/players/S/StewAr00.htm'

#has a stat kick and punt returns and nfl stats inverted
dri_archer_nfl_url = 'https://www.pro-football-reference.com/players/A/ArchDr00.htm'

url_nfl_sample = odell_beckham_jr_nfl_url
url = url_nfl_sample
html =urlopen(url_nfl_sample)
soup = BeautifulSoup(html,'html.parser')

table=nflstats(soup)

display(table)

#row_nfl(table)

## Code: create the table of draft without merging with tables of cfb and nfl

#url_wr_draft = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=0'
url_wr_draft = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=300'
html =urlopen(url_wr_draft)
soup = BeautifulSoup(html,'html.parser')

table0=tabledraft(soup)

table1 = cfb_url_fill_in(table0)
table1 = expend_draft_table(table1)

## Code: merging data from Julio Jones' cfb/nfl to draft table

#index 255
table1=table_draft_1
julio_jones_cfb_url = table1.loc[table1.loc[:,'player'] == 'Julio Jones']['cfb url'].item()
print(julio_jones_cfb_url)
julio_jones_nfl_url = table1.loc[table1.loc[:,'player'] == 'Julio Jones']['nfl url'].item()
print(julio_jones_nfl_url)

url_cfb_sample = julio_jones_cfb_url
html = urlopen(url_cfb_sample)
soup = BeautifulSoup(html,'html.parser')

table2=collegestats(soup)
display(table2)
print(row_cfb(table2))

url_nfl_sample = julio_jones_nfl_url
html =urlopen(url_nfl_sample)
soup = BeautifulSoup(html,'html.parser')

table3=nflstats(soup)
display(table3)
print(row_nfl(table3))

#table2.iloc[-2][0:4].tolist()+ [str(table2['games'].apply(pd.to_numeric).iloc[:-1].sum())] + table2.iloc[-1][5:].tolist()
#table3.iloc[-2][0:4].tolist()+ table3.iloc[-1][4:].tolist()


#table1.loc[255,24:72] = table2.iloc[-2][0:4].tolist()+ [str(table2['games'].apply(pd.to_numeric).iloc[:-1].sum())] + table2.iloc[-1][5:].tolist()+table3.iloc[-2][0:4].tolist()+ table3.iloc[-1][4:].tolist()

## Uselul commands
Do not run unless initialized

os.getcwd()

url_wr_draft_1 = 'https://www.pro-football-reference.com/play-index/draft-finder.cgi?request=1&year_min=1936&year_max=2019&type=&round_min=1&round_max=30&slot_min=1&slot_max=500&league_id=&team_id=&pos%5B%5D=WR&college_id=all&conference=any&show=all&offset=0'
html =urlopen(url_wr_draft_1)
soup = BeautifulSoup(html,'html.parser')

wb.open_new_tab(url_wr_draft_1)
table_0 = soup.find_all('table')[0]
rowx= table_0.find_all('tr')[4]
columnx = rowx.find_all('td')[3]
linkx = columnx.find_all('a')[0]
print('https://www.pro-football-reference.com'+linkx.get('href'))

#list of added columns
#['cfb '+ x for x in table2.columns.tolist()] + ['nfl '+ x for x in table3.columns.tolist()]
