==> Capstone_1 (old)/2) Collecting_Data_TableCombine.ipynb <==
### Combine Notebook:
The notebook filters out WR between 2000 and 2011. Still, the file we use to merge with the draft table might be the original, as we need to account for players who are not WR by the time of the combine, and yet get drafted such as

import pandas as pd

#import table
table0 = pd.read_csv('data_sets/combine_data_since_2000.csv')

print(table0['Year'].unique())
print(table0['Pos'].unique())

#slice dataframe and export
table1 = table0[(table0['Pos']=='WR') & (table0['Year'] < 2012) & (table0['Year']>=2000)].reset_index(drop=True)
table1.to_csv(r'data_sets/wr_combine.csv',index=False)
