==> Capstone_1 (old)/4) Cleaning_Data.ipynb <==
# Cleaning Data:

The objective of this notebook is to "clean" the data. The process goes in multiple steps:
- re-organizing data in a more natural and explicit way
- replacing error messages with np.nan entries for missing data
- catch left over errors in the nfl data via an automated function or manually
- carefully convert all numerical values to floats (with a few exception such as years and draft positions)
- correct and replace improper (not missing) cfb data
- finally, fill in np.nan with 0 when we're certain this is the correct value (when a player doesn't have any rushing stats for example)

Once done we export our clean table for usage in the Exploratory notebook

import pandas as pd
import numpy as np
from urllib.request import urlopen
from bs4 import BeautifulSoup
from collections import Counter
from selenium import webdriver
import time

pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', 100)
pd.options.display.max_colwidth = 100

list_of_college = ['school','conference','class','pos','games','receptions',
                   'yards', 'average','td','attemps rushing','yards rushing',
                   'avg rushing','td rushing','scrimmages','yards total','avg total','td total']

tdc = pd.read_csv('data_sets/wr_draft_combine.csv')

### Re-organizing and re-naming the columns in proper order

tdc = tdc[['year', 'round', 'pick', 'player', 'pos', 'draft age',
       'team', 'entry year', 'last year', '1st team pro select', 'pro select',
       'weighted career av', 'years as primary starter', 'games',
       'games started', 'rushing attempts', 'rushing yards', 'rushing td',
       'receiving attemps', 'receiving yards', 'receiving td',
        'nfl url','nfl age', 'nfl team', 'nfl pos', 'nfl no', 'nfl game',
       'nfl game started', 'nfl target', 'nfl receptions', 'nfl yards',
       'nfl y/r', 'nfl td', 'nfl first downs', 'nfl longest rec',
       'nfl rec per game', 'nfl yards per game', 'nfl catch ratio',
       'nfl yards per target', 'nfl rushes', 'nfl rush yards', 'nfl rush td',
       'nfl first downs rush', 'nfl longest rush',
       'nfl rush yards per attempt', 'nfl rush yards per game',
       'nfl rush attempt per games', 'nfl total touches',
       'nfl yards per touch', 'nfl yards from scrimmage', 'nfl total td',
       'nfl fumbles', 'nfl av', 'nfl table type',
        'college', 'cfb url', 'cfb school', 'cfb conference', 'cfb class', 'cfb pos',
       'cfb games', 'cfb receptions', 'cfb yards', 'cfb average', 'cfb td',
       'cfb attemps rushing', 'cfb yards rushing', 'cfb avg rushing',
       'cfb td rushing', 'cfb scrimmages', 'cfb yards total', 'cfb avg total',
       'cfb td total',
       'combine player', 'combine pos', 'combine ht', 'combine wt',
       'combine forty', 'combine vertical', 'combine benchreps',
       'combine broadjump', 'combine cone', 'combine shuttle', 'combine year',
       'combine pfr_id', 'combine av', 'combine team', 'combine round',
       'combine pick',
        'nfl method','cfb method', 'combine method']]


tdc = tdc.rename(columns={'weighted career av':'years as primary starter',
                    'years as primary starter':'weighted career average grade',
                    'rushing attempts':'rush attempts',
                    'rushing yards':'rush yards',
                    'rushing td':'rush td',
                    'receiving attemps':'rec attempts',
                    'receiving yards':'rec yards',
                    'receiving td':'rec td',
                    'nfl target':'nfl pass targets',
                    'nfl yards':'nfl rec yards',
                    'nfl y/r':'nfl rec yards per reception',
                    'nfl yards per target':'nfl rec yards per target',
                    'nfl td':'nfl rec td',
                    'nfl first downs':'nfl 1st downs from reception',
                    'nfl longest rec':'nfl longest reception',
                    'nfl rec per game':'nfl receptions per game',
                    'nfl yards per game':'nfl rec yards per game',
                   'nfl rushes':'nfl rush attempts',
                    'nfl first downs rush':'nfl 1st downs from rush',
                   'nfl rush attempt per games':'nfl rush attempts per game',
                    'nfl total touches':'nfl touches',
                    'nfl total td':'nfl td',
                   'nfl av': 'nfl average grade',
                   'cfb yards':'cfb rec yards',
                   'cfb average':'cfb rec yards per reception',
                   'cfb td':'cfb rec td',
                   'cfb attemps rushing':'cfb rush attempts',
                   'cfb yards rushing':'cfb rush yards',
                   'cfb avg rushing':'cfb rush yards per attempt',
                   'cfb td rushing':'cfb rush td',
                   'cfb yards total':'cfb yards',
                   'cfb avg total':'cfb yards from scrimmage',
                   'cfb td total':'cfb td'})

### Cleaning columns: replace error messages with np.nan
Converting obvious missing data to np.nan, such as rows with no cfb stats (eg broken cfb link), or rows with no nfl stats (eg broken nfl link) 

cond = [str.split(x)[0] == 'fail' for x in tdc['cfb method']]
tdc.loc[cond,'cfb school':'cfb td'] = [np.nan]*17

cond = [str.split(x)[0] == 'fail' for x in tdc['combine method']]

cond= [str.split(row['nfl method'])[0] == 'fail' and str.split(row['nfl team'])[0] == 'fail' for i,row in tdc.iterrows()]
tdc.loc[cond,'nfl age':'nfl table type'] = [np.nan]*32

### Catching some NFL errors
Some tables have rushing and receiving still switched up. We find those if the value of 'nfl catch ratio' doesn't end with '%'

error_list = []
col = 'nfl catch ratio'
for i in tdc[col].index:
    if isinstance(tdc.loc[i,col], str):
        if tdc.loc[i,col][-1] != '%':
            error_list.append(i)
            
#display(tdc.loc[[0]+error_list])
#[print(x) for x in tdc.loc[[0]+error_list]['nfl url']]

i_col = tdc.columns.get_loc('nfl pass targets')
lr = 8
lw = 11

def switch_rushing_receiving (table, error_list): #switch rush data and receiving data if needed
    for i in error_list:
        rush = table.iloc[i,i_col:i_col+lr].tolist()
        receiving = table.iloc[i,i_col+lr:i_col+lr+lw].tolist()
        table.iloc[i,i_col:i_col+lw] = receiving
        table.iloc[i,i_col+lw:i_col+lw+lr] = rush
    return table

tdc = switch_rushing_receiving(tdc,error_list)
#tdc.loc[error_list]

### Sets of manual NFL rectifications
Let's see if we can catch some left over errors and edit them manually. For example, some players had numbers in the column 'nfl team', revealing the data wasn't properly treated. We take care of it here

#tdc['nfl team'].value_counts()
rectif_1 = tdc.index[(tdc['nfl team']=='24') | (tdc['nfl team']=='27')].tolist()
#tdc.loc[rectif_1]
#[print(x) for x in tdc.loc[rectif_1]['nfl url']]

for i in rectif_1:
    tdc.loc[i,'nfl team':'nfl average grade'] =[tdc.loc[i,'nfl pos']]+[np.nan]+[tdc.loc[i,'nfl game']]+[np.nan]*27

tdc.to_csv('data_sets/wr_tdc_clean.csv', index=False)

### Converting numerical strings to floats carefully
Starting with the column of nfl catch ratio, which is a string with '%' at the end

tdc = pd.read_csv('data_sets/wr_tdc_clean.csv')

col = 'nfl catch ratio'
for i in tdc[col].index:
    if isinstance(tdc.loc[i,col], str):
        s = tdc.loc[i,col][:-1]
        tdc.loc[i,col] = float(s)/100
        
tdc[col] = tdc[col].astype(float)

Let's now display the type of all columns

for i, col in enumerate(tdc.columns.tolist()):
    print(i, col, tdc[col].dtype)

tdc.to_csv('data_sets/wr_tdc_clean.csv', index=False)

### Improper cfb data

It seems like some of the cfb rows have not been properly scrapped. For example, some of the players have outrageous rec td in cfb

tdc = pd.read_csv('data_sets/wr_tdc_clean.csv')

Let's look at all the cfb table captions we scraped and hope to find an error

def find_caption_from_url(url):
    soup = BeautifulSoup(urlopen(url),'html.parser')
    table_0 = soup.find_all('table')[0]
    try:
        x=table_0.find_all('caption')[0].get_text()
    except:
        x='no caption'
    return x

caption_list=[]
for i, row in tdc.iterrows():
    url = row['cfb url']
    try:
        caption_list.append(find_caption_from_url(url))
    except:
        caption_list.append(np.nan)
        
caption_df = pd.DataFrame({'captions':caption_list})

caption_df = pd.read_csv('dfcap.csv')

caption_df['captions'].value_counts()

There is in fact multiple tables with different captions that expected:

- For those with "Rushing ..." we simply have to switch the order of the receiving and rushing stats
- For the others, we need to do our advanced "web scrapping" which allows us to find all the tables within one page. As we find the proper caption, we can extract the data, without forgetting to convert to floats when possible
- Finally, for this last category, there is a chance that we find the table 'Rushing & Receiving Table' rather than its proper counterpart, so we need to do the inverting of those as well

ind = caption_df.index[caption_df['captions'] == 'Rushing & Receiving Table'].tolist()

def switch_rushing_receiving1 (tdc,ind): #switch rush data and receiving data if needed
    for i in ind:
        rush = tdc.iloc[i,61:65]
        receiving = tdc.iloc[i,65:69]

        tdc.iloc[i,61:69] = receiving.tolist() +  rush.tolist()
    return tdc

tdc = switch_rushing_receiving1(tdc, ind)

ind = caption_df.index[(caption_df['captions'] == 'Passing Table')|
                      (caption_df['captions'] == 'Kick & Punt Returns Table')|
                      (caption_df['captions'] == 'Defense & Fumbles Table')|
                      (caption_df['captions'] == 'Punt & Kick Returns Table')].tolist()


def generate_all_nfl_tables(url): #use web.driver chrome to find hidden tables
    driver = webdriver.Chrome('/Users/adriensaremi/Documents/School and Employment/SpringBoard Course/Springboard/Capstone_1/data_sets/chromedriver')
    driver.get(url)
    soup = BeautifulSoup(driver.page_source,'lxml')
    driver.quit()
    return soup.find_all('table')


def generate_proper_cfb_row(table_list): #return the proper row from the table list
    cl = []
    row = [0]
    for x in table_list:
        try:
            caption = x.find_all('caption')[0].get_text()
        except:
            caption = 'no'           
        if caption == 'Receiving & Rushing Table':
            table_cfb = collegestats(x)
            row = row_cfb(table_cfb)
            break
        elif caption == 'Rushing & Receiving Table':
            table_cfb = collegestats(x)
            row = row_cfb(table_cfb)
            row = switch_rushing_receiving2(row)
            break
        else:
            pass

    return row


def collegestats(table_0): #provide cfb stats from table
    table = pd.DataFrame(columns=list_of_college,
                         index=range(0,15)) #create a new blank table
    
    row_marker = 0
    for row in table_0.find_all('tr'):
        column_marker = 0
        columns = row.find_all('td')
        for column in columns:
            table.iat[row_marker,column_marker] = column.get_text()
            column_marker += 1
        row_marker += 1
    
    table.dropna(subset=['school'],inplace=True)  

    return table.reset_index(drop = True)  

def overall_cfb_index(t): #which row is the overall stats. Needed for players with multiple teams
    try: 
        list = [x.lower() for x in t['school'].tolist()]
        i = list.index('overall')
    except:
        i = len(t)-1
    return i

def row_cfb(t):#picks the overall row of the college tabler and adds data from the year right before (team, age, pos)
    i = overall_cfb_index(t)
    return t.loc[i-1][0:4].tolist() + [str(t['games'].apply(pd.to_numeric).iloc[:i].sum())] + t.iloc[i][5:].tolist()

def switch_rushing_receiving2 (l1): #switch rush data and receiving data if needed
    l2 = l1
    rush = l1[5:9]
    receiving = l1[9:13]
    l2[5:13] = receiving + rush
    return l2

def main(tdc, ind): #use a for loop to do it all
    row_list = []
    for i in ind:
        table_list = generate_all_nfl_tables(tdc.loc[i,'cfb url'])
        time.sleep(5)
        row = generate_proper_cfb_row(table_list)
        row_list.append(row)                  
    
    return row_list

#rows = main(tdc, ind)
#pd.DataFrame(rows).to_csv('rows_of_search.csv', index = False)

def mainp (tdc,ind,rows):
    r = 0
    for i in ind:
        try:
            tdc.loc[i,'cfb school':'cfb td'] = rows[r][0:4] + [float(x) for x in rows[r][4:]]
        except:
            tdc.loc[i,'cfb school':'cfb td'] = tdc.loc[i,'cfb school':'cfb pos'].tolist() + [np.nan]*13
        r = r +1
        
    return tdc

#tdcc = tdc.copy()
#tdcp = mainp(tdcc,ind, rows)
tdc = mainp(tdc,ind,rows)

tdc.to_csv('data_sets/wr_tdc_clean.csv', index=False)

### Adding zero's

Let's replace the np.nan or '' entries with zero's as long as we're sure of

tdc = pd.read_csv('data_sets/wr_tdc_clean.csv')

tdc.loc[:,'weighted career average grade':'rec td'] = tdc.loc[:,'weighted career average grade':'rec td'].fillna(0)

tdc.iloc[2]['year':'nfl game']

for i, row in tdc.iterrows():
    if row['cfb method'] == 'success':
        tdc.loc[i,'cfb games':'cfb td'] = tdc.loc[i,'cfb games':'cfb td'].fillna(0).values.tolist()

tdc.iloc[8]['cfb games':'cfb td']

for i, row in tdc.iterrows():
    if row['nfl method'].split()[0] == 'success':
        tdc.loc[i,'nfl game':'nfl fumbles'] = tdc.loc[i,'nfl game':'nfl fumbles'].fillna(0).values.tolist()

tdc.iloc[53]['nfl game':'nfl fumbles']

tdc.to_csv('data_sets/wr_tdc_clean.csv', index=False)
